# tika-server-old/tika-server-spring/src/main/resources/api/tika-server-openapi.yaml
...
openapi: 3.0.2
...
info:
  contact:
    email: dev@tika.apache.org
  description: The Tika server implements [JAX-RS](http://en.wikipedia.org/wiki/JAX-RS) (Java API for RESTful Web Services) to provide web services according to the Representational State Transfer (REST) architectural style. This facilitates a wide varity oif operations and flexibility with regards to both client and server implementations. The officially supported Tika server implementation is packaged using the OpenAPI [jaxrs-cxf generator](https://openapi-generator.tech/docs/generators/jaxrs-cxf]. This work was tracked through [TIKA-3082](https://issues.apache.org/jira/browse/TIKA-3082).
    <b>N.B.</b> the OpenAPI version always tracks the underlying Tika version to remove uncertainty about which version of Tika is running within the server.
  license:
    name: Apache v2
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Tika JAX-RS Server
  version: 4.0.0-SNAPSHOT
servers:
  - description: Local development server
    url: 'https://localhost:9998'
...
paths:
...
  /tika:
    get:
        description: 'HTTP GET returns a greeting stating the server is up. Followed by a PUT request to extract text.'
        operationId: get_tika
        responses:
          '200':
            content:
              text/plain:
                schema:
                  example: 'This is Tika Server (Apache Tika 4.0.0-SNAPSHOT). Please PUT'
                  type: string
            description: If successful, this operation returns HTTP status code 200, with a greeting to indicate the server is up and you may PUT a file.
            headers:
              Content-Type:
                schema:
                  example: text/plain
                  default: text/plain
                  type: string
                description: Content-type served to client.
          '500':
            description: An error occurred processing the call.
        summary: GET returns a greeting stating the server is up.
        tags:
        - Tika Resource
    put:
        description: 'HTTP PUTs a document to the /tika service and you get back the extracted text.'
        operationId: put_tika
        responses:
          '200':
            content:
              text/plain:
                schema:
                  type: string
            description: If successful, this operation returns HTTP status code 200, with the extracted text.
            headers:
              Content-Type:
                schema:
                  example: text/plain
                  default: text/plain
                  type: string
                description: Content-type served to client.
          '500':
            description: An error occurred processing the call.
        summary: PUT document to extract text.
        tags:
        - Tika Resource
  /tika/form:
    post:
      description: 'POST a document via multipart/form-data to the /tika service for text extraction.'
      operationId: post_tika_form
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: true
      responses:
        '200':
          content:
            text/plain:
              schema:
                type: string
            text/html:
              schema:
                type: string
            text/xml:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
          description: If successful, returns the extracted content in the requested format.
        '500':
          description: An error occurred processing the call.
      summary: 'POST form for text extraction.'
      tags:
      - Tika Resource
  /tika/main:
    put:
      description: 'PUT a document and use Boilerpipe to extract the main content (equivalent to text-main in tika-app).'
      operationId: put_tika_main
      responses:
        '200':
          content:
            text/plain:
              schema:
                type: string
          description: If successful, this operation returns HTTP status code 200, with the main extracted text.
        '500':
          description: An error occurred processing the call.
      summary: 'PUT document to extract main content using Boilerpipe.'
      tags:
      - Tika Resource
  /tika/form/main:
    post:
      description: 'POST a document via multipart/form-data to extract main content using Boilerpipe.'
      operationId: post_tika_form_main
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: true
      responses:
        '200':
          content:
            text/plain:
              schema:
                type: string
          description: If successful, returns the main extracted text.
        '500':
          description: An error occurred processing the call.
      summary: 'POST form for main content extraction.'
      tags:
      - Tika Resource
  /tika/{handler}:
    put:
      description: 'PUT a document with a specific content handler type.'
      parameters:
      - description: Handler type for content (xml, text, html, ignore)
        explode: false
        in: path
        name: handler
        required: true
        schema:
          example: 'xml'
          type: string
          enum: [xml, text, html, ignore]
        style: simple
      operationId: put_tika_handler
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
          description: If successful, returns the processed content with metadata.
        '500':
          description: An error occurred processing the call.
      summary: 'PUT document with specific handler type.'
      tags:
      - Tika Resource
...
components:
  schemas:
...
    Metadata:
      description: "The typical Tika multi-valued metadata key-value map."
      type: object
      additionalProperties:
        type: string
...